# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  devices.Device:
    properties:
      agent_version:
        type: string
      cpu_cores:
        type: integer
      cpu_model:
        type: string
      created_at:
        type: string
      customer_id:
        type: string
      customer_name:
        type: string
      disk_total_mb:
        type: integer
      hostname:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_online:
        type: boolean
      last_heartbeat:
        type: string
      mssql_status:
        type: string
      notes:
        type: string
      os_version:
        type: string
      pos_process_status:
        type: string
      ram_total_mb:
        type: integer
      rustdesk_id:
        type: string
      rustdesk_password:
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  devices.DeviceCreateRequest:
    properties:
      customer_id:
        type: string
      hostname:
        type: string
      notes:
        type: string
      os_version:
        type: string
      rustdesk_id:
        type: string
    required:
    - hostname
    type: object
  devices.DeviceListResponse:
    properties:
      devices:
        items:
          $ref: '#/definitions/devices.Device'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  devices.DeviceUpdateRequest:
    properties:
      customer_id:
        type: string
      hostname:
        type: string
      mssql_status:
        type: string
      notes:
        type: string
      pos_process_status:
        type: string
      rustdesk_id:
        type: string
      rustdesk_password:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  devices.HeartbeatRequest:
    properties:
      agent_version:
        type: string
      cpu_cores:
        type: integer
      cpu_model:
        type: string
      cpu_percent:
        type: number
      disk_percent:
        type: number
      disk_total_mb:
        type: integer
      disk_used_mb:
        type: integer
      hostname:
        type: string
      mssql_running:
        type: boolean
      os_version:
        type: string
      pos_running:
        type: boolean
      ram_percent:
        type: number
      ram_total_mb:
        type: integer
      ram_used_mb:
        type: integer
      rustdesk_id:
        type: string
      uptime_seconds:
        type: integer
    required:
    - hostname
    type: object
...
  tickets.AddCommentRequest:
    properties:
      content:
        type: string
      is_internal:
        type: boolean
    required:
    - content
    type: object
  tickets.CreateTicketRequest:
    properties:
      alert_id:
        type: string
      assigned_to:
        type: string
      customer_id:
        type: string
      description:
        type: string
      device_id:
        type: string
      priority:
        type: string
      source:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  tickets.Ticket:
    properties:
      alert_id:
        type: string
      assigned_name:
        type: string
      assigned_to:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      created_name:
        type: string
      customer_id:
        type: string
      customer_name:
        type: string
      description:
        type: string
      device_id:
        type: string
      device_name:
        description: Joined
        type: string
      id:
        type: string
      priority:
        type: string
      resolved_at:
        type: string
      source:
        type: string
      status:
        type: string
      ticket_number:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  tickets.TicketComment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_internal:
        type: boolean
      ticket_id:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  tickets.TicketListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      tickets:
        items:
          $ref: '#/definitions/tickets.Ticket'
        type: array
      total:
        type: integer
    type: object
  tickets.UpdateTicketRequest:
    properties:
      assigned_to:
        type: string
      description:
        type: string
      priority:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Remote Monitoring & Management Platform API
  title: RMM Platform API
  version: "1.0"
paths:
...
  /devices:
    get:
      description: Get paginated list of devices
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Search by hostname or customer
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devices.DeviceListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Register a new device
      parameters:
      - description: Device data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devices.DeviceCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/devices.Device'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create device
      tags:
      - devices
  /devices/{id}:
    delete:
      description: Soft-delete a device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete device
      tags:
      - devices
    get:
      description: Get detailed device information
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devices.Device'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get device by ID
      tags:
      - devices
    put:
      consumes:
      - application/json
      description: Update device information
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Device update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devices.DeviceUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devices.Device'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update device
      tags:
      - devices
...
  /devices/heartbeat:
    post:
      consumes:
      - application/json
      description: Process agent heartbeat with system metrics
      parameters:
      - description: Heartbeat data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devices.HeartbeatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devices.Device'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Receive device heartbeat
      tags:
      - devices
  /tickets:
    get:
      description: Get paginated list of support tickets
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by device ID
        in: query
        name: device_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tickets.TicketListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List tickets
      tags:
      - tickets
    post:
      consumes:
      - application/json
      description: Create a new support ticket
      parameters:
      - description: Ticket data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tickets.CreateTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tickets.Ticket'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create ticket
      tags:
      - tickets
  /tickets/{id}:
    get:
      description: Get ticket details with comments
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get ticket by ID
      tags:
      - tickets
    put:
      consumes:
      - application/json
      description: Update ticket status or details
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Ticket update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tickets.UpdateTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tickets.Ticket'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update ticket
      tags:
      - tickets
  /tickets/{id}/comments:
    post:
      consumes:
      - application/json
      description: Add a comment to a support ticket
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tickets.AddCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tickets.TicketComment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add ticket comment
      tags:
      - tickets
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"